+++ src/dictionaries.h
@@ -1,0 +1,2 @@
+SEXP zstd_dict_id_(SEXP src_);
+