+++ src/utils.h
@@ -1,1 +1,2 @@
 unsigned char *read_file(const char *filename, size_t *src_size); 
+unsigned char *read_partial_file(const char *filename, size_t max_bytes, size_t *src_size);